+++ src/frontend/data-dog-browser-log-sink.tsx
